LKE-12088 data model for query builder (#758)
